Cast one of the operands of this multiplication operation to a 'double'

+++ src/main/java/NewReleasePrice.java
@@ -4,7 +4,7 @@ class NewReleasePrice extends Price {
 		return Movie.NEW_RELEASE;
 	}
 	double getCharge(int daysRented) {
-		return daysRented * 3;
+		return (double) daysRented * 3;
 	}
 	public int getFrequentRenterPoints(int daysRented) {
 		return (daysRented > 1) ? 2 : 1;
